requirements.txt
langchain
langchain-community
langchain-openai
sentence-transformers
chromadb
python-dotenv
openai
tiktoken
streamlit